refactor(web): compute code text without a render-phase ref (#888)
PreBlock derives codeText as a plain const via the shared helper;
the ref read/write during render is gone.
