+++ ornn-web/src/components/docs/DocsMarkdownComponents.tsx
@@ -13,17 +13,9 @@
  * @module components/docs/DocsMarkdownComponents
  */
 
-import { useRef, useState } from "react";
+import { useState } from "react";
 import { MermaidBlock } from "./DocsMermaid";
-
-export function slugify(text: string): string {
-  return text
-    .toLowerCase()
-    .replace(/[^\w\s\u4e00-\u9fff-]/g, "")
-    .replace(/\s+/g, "-")
-    .replace(/-+/g, "-")
-    .replace(/^-|-$/g, "");
-}
+import { slugify, extractTextFromChildren } from "./DocsMarkdownComponents.helpers";
 
 function CopyButton({ text }: { text: string }) {
   const [copied, setCopied] = useState(false);
@@ -51,25 +43,14 @@ function CopyButton({ text }: { text: string }) {
 
 /** Wraps <pre> blocks with a relative container and copy button outside the scrollable area */
 function PreBlock({ children, ...props }: React.HTMLAttributes<HTMLPreElement>) {
-  const codeText = useRef("");
-
-  // Extract text content from children (the <code> element inside <pre>)
-  const extractText = (node: React.ReactNode): string => {
-    if (typeof node === "string") return node;
-    if (typeof node === "number") return String(node);
-    if (Array.isArray(node)) return node.map(extractText).join("");
-    if (node && typeof node === "object" && "props" in node) {
-      const el = node as { props?: { children?: React.ReactNode } };
-      return extractText(el.props?.children);
-    }
-    return "";
-  };
-
-  codeText.current = extractText(children).replace(/\n$/, "");
+  // Pure derivation from props — no ref needed. The copy text is the
+  // flattened text content of the <code> child with a trailing newline
+  // stripped (markdown code fences carry one).
+  const codeText = extractTextFromChildren(children).replace(/\n$/, "");
 
   return (
     <div className="relative">
-      <CopyButton text={codeText.current} />
+      <CopyButton text={codeText} />
       <pre {...props}>{children}</pre>
     </div>
   );
@@ -97,17 +78,6 @@ function CodeBlock({
   );
 }
 
-function extractTextFromChildren(children: React.ReactNode): string {
-  if (typeof children === "string") return children;
-  if (typeof children === "number") return String(children);
-  if (Array.isArray(children)) return children.map(extractTextFromChildren).join("");
-  if (children && typeof children === "object" && "props" in children) {
-    const el = children as { props?: { children?: React.ReactNode } };
-    return extractTextFromChildren(el.props?.children);
-  }
-  return "";
-}
-
 function H1({ children, ...props }: React.HTMLAttributes<HTMLHeadingElement>) {
   const id = slugify(extractTextFromChildren(children));
   return <h1 id={id} {...props}>{children}</h1>;
@@ -125,15 +95,8 @@ function H4({ children, ...props }: React.HTMLAttributes<HTMLHeadingElement>) {
   return <h4 id={id} {...props}>{children}</h4>;
 }
 
-/**
- * Component map to pass into `<ReactMarkdown components={...}>`.
- * Renames `pre`/`code` and injects slug IDs on h1-h4.
- */
-export const markdownComponents = {
-  pre: PreBlock,
-  code: CodeBlock,
-  h1: H1,
-  h2: H2,
-  h3: H3,
-  h4: H4,
-};
+// The `markdownComponents` map (assembled from these components) lives
+// in the sibling `DocsMarkdownComponents.map.ts` so this file only
+// exports components — keeping react-refresh's Fast Refresh boundary
+// intact (#888).
+export { PreBlock, CodeBlock, H1, H2, H3, H4 };
